requirements.txt
PyOpenGL==3.1.6
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
